Add Colorful Text, fix killfeed, fix loadingscreen, UI
