--- Assets/Kilfeed.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ada566588fabf254cbb97a0222c6319a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
