+++ Assets/Material/Player1/#AAA7F7.mat
@@ -108,8 +108,8 @@ Material:
     - _WorkflowMode: 1
     - _ZWrite: 1
     m_Colors:
-    - _BaseColor: {r: 0.49803922, g: 0.68235296, b: 0.13725491, a: 1}
-    - _Color: {r: 0.49803922, g: 0.68235296, b: 0.13725491, a: 1}
+    - _BaseColor: {r: 0.99215686, g: 0.68235296, b: 0.79607844, a: 1}
+    - _Color: {r: 0.99215686, g: 0.68235296, b: 0.79607844, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.2, g: 0.2, b: 0.2, a: 1}
   m_BuildTextureStacks: []
