+++ Assets/Scripts/Result.cs
@@ -12,11 +12,6 @@ public class Result : MonoBehaviour
 
     [SerializeField]
     ScoreManager scoreManager;
-
-    /*[SerializeField]
-    private TextMeshProUGUI player1ScoreText;
-    [SerializeField]
-    private TextMeshProUGUI player2ScoreText;*/
     [SerializeField]
     private TextMeshProUGUI[] playerScoreText;
     [SerializeField]
@@ -25,12 +20,16 @@ public class Result : MonoBehaviour
     void OnEnable()
     {
         MatchHistory history = new MatchHistory();
+
         MatchRecordManager matchRecordManager = new MatchRecordManager();
+
         for(int i = 0; i < playerScoreText.Length; i++)
         {
             playerScoreText[i].text = "Player " + (i+1) + " score : " + scoreManager.GetPlayerScore(i).ToString();
         }
-        
+
+        string winner = null;
+
         int highScore = scoreManager.GetPlayerScore().Max();
         int highScoreIndex = scoreManager.GetPlayerScore().ToList().IndexOf(highScore);
 
@@ -47,30 +46,32 @@ public class Result : MonoBehaviour
         if (sameScore == 1)
         {
             playerWonText.text = "Player "+ (highScoreIndex + 1) +" Won";
-
-            int[] index = scoreManager.GetPlayerScore();
-            string[] dataPlayer = new string[index.Length];
-
-            string[] playerscolor = new string[scoreManager.GetPlayerScore().Length];
-            string[] playersID = new string[scoreManager.GetPlayerScore().Length];
-
-            for(int i=0; i<index.Length; i++)
-            {
-                dataPlayer[i] = "Player" + (i + 1);
-                Debug.Log(dataPlayer[i]);
-            }
-            for(int i=0; i<playerscolor.Length; i++)
-            {
-                playersID[i] = "Player" + (i + 1);
-                playerscolor[i] = PlayerPrefs.GetString(playersID[i] + "Color");
-            }
-            history.PlayerRecord(dataPlayer, "Player"+ (highScoreIndex + 1));
-            matchRecordManager.RecordMatch(playersID, scoreManager.GetPlayerScore(), playerscolor);
+            winner = "Player" + (highScoreIndex + 1);
         }
         else
         {
             playerWonText.text = "Draw !";
         }
+
+        int[] index = scoreManager.GetPlayerScore();
+        string[] dataPlayer = new string[index.Length];
+
+        string[] playerscolor = new string[scoreManager.GetPlayerScore().Length];
+        string[] playersID = new string[scoreManager.GetPlayerScore().Length];
+
+        for (int i = 0; i < index.Length; i++)
+        {
+            dataPlayer[i] = "Player" + (i + 1);
+            Debug.Log(dataPlayer[i]);
+        }
+        for (int i = 0; i < playerscolor.Length; i++)
+        {
+            playersID[i] = "Player" + (i + 1);
+            playerscolor[i] = PlayerPrefs.GetString(playersID[i] + "Color");
+        }
+
+        history.PlayerRecord(dataPlayer, winner);
+        matchRecordManager.RecordMatch(playersID, scoreManager.GetPlayerScore(), playerscolor);
 
     }
 }
