+++ Assets/SkyBoxes/Daybreak/Daybreak.mat
@@ -98,7 +98,7 @@ Material:
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
-    - _Rotation: 90.120964
+    - _Rotation: 126.1842
     - _SrcBlend: 1
     - _SunDisk: 2
     - _SunSize: 0.732
