+++ Assets/SkyBoxes/Midday/Midday.mat
@@ -98,7 +98,7 @@ Material:
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
-    - _Rotation: 11.366152
+    - _Rotation: 63.0921
     - _SrcBlend: 1
     - _SunDisk: 2
     - _SunSize: 0.732
