+++ Assets/SkyBoxes/Sunset/Sunset.mat
@@ -98,7 +98,7 @@ Material:
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
-    - _Rotation: 368.08618
+    - _Rotation: 85.17947
     - _SrcBlend: 1
     - _SunDisk: 2
     - _SunSize: 0.732
